% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vaccHamilton.R
\docType{data}
\name{data_da_2016}
\alias{data_da_2016}
\title{Hamilton CMA Dissemination Areas with Selected Census Information.}
\format{
A simple features data frame with 1,199 rows and 25 variables:
\describe{
  \item{Population}{Population in the DA}
  \item{Households}{Number of households in DA}
  \item{GeoUID}{Unique identifier of the DA}
  \item{Type}{Geography type}
  \item{Shape Area}{Area of the shape}
  \item{Dwellings}{Number of dwellings in DA}
  \item{CSD_UID}{Census Subdivision (CSD) unique identifier}
  \item{CT_UID}{Census Tract (CT) unique identifier}
  \item{CMA_UID}{Census Metropolitan Area (CMA) unique identifier}
  \item{Region Name}{Name of the region within the CMA}
  \item{Area (sq km)}{Area of the DA in square kilometers}
  \item{v_CA16_190: 50 to 54 years}{Population age 50-54 in the DA}
  \item{v_CA16_208: 55 to 59 years}{Population age 55-59 in the DA}
  \item{v_CA16_226: 60 to 64 years}{Population age 60-64 in the DA}
  \item{v_CA16_247: 65 to 69 years}{Population age 65-69 in the DA}
  \item{v_CA16_265: 70 to 74 years}{Population age 70-74 in the DA}
  \item{v_CA16_283: 75 to 79 years}{Population age 75-79 in the DA}
  \item{v_CA16_301: 80 to 84 years}{Population age 80-84 in the DA}
  \item{v_CA16_319: 85 years and over}{Population 85 years and older in the DA}
  \item{v_CA16_2397: Median total income of households in 2015 ($)}{Median income of households in the DA}
  \item{v_CA16_5795: Car, truck, van - as a driver}{Number of people whose main mode of transportation to work in the DA is as drivers}
  \item{v_CA16_5801: Public transit}{Number of people whose main mode of transportation to work in the DA is transit}
  \item{v_CA16_5804: Walked}{Number of people whose main mode of transportation to work in the DA is walking}
  \item{geometry}{Geometry of boundaries}
}
}
\source{
Statistics Canada
}
\usage{
data(data_da_2016)
}
\description{
Boundaries of the Dissemination Areas in Hamilton CMA.
}
\keyword{Canada}
\keyword{Census}
\keyword{datasets}
\keyword{transportation}
