% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vaccHamilton.R
\docType{data}
\name{pharmacy_locations}
\alias{pharmacy_locations}
\title{Location of pharmacies offering COVID-19 vaccinations.}
\format{
A simple features data frame with 20 rows and 3 variables:
\describe{
  \item{ID}{Unique identifier of pharmacy}
  \item{Location}{Address of pharmacy}
  \item{Type}{Type of location; "Pilot": those in the provincial pilot program; "Candidate": additional candidate location; "Existing for 70+": other vaccination sites available for people 70 y.o. and older}
  \item{geometry}{Geometry of points}
}
}
\source{
Statistics Canada
}
\usage{
data(pharmacy_locations)
}
\description{
Location of pharmacies offering COVID-19 vaccinations in the City of Hamilton.
}
\keyword{datasets}
\keyword{land}
\keyword{transportation}
\keyword{use}
